+++ crates/digital_twin_glue/src/physics/solver.rs
@@ -3,8 +3,8 @@ use crate::geometry::mesh::Mesh;
 use nalgebra::Vector3;
 
 pub struct SolverOutput {
-    pub force: Vector3<f32>,
-    pub torque: Vector3<f32>,
+    pub force: Vector3<f64>,
+    pub torque: Vector3<f64>,
 }
 
 pub trait AeroSolver {
